+++ packages/taro-components/src/components/block/index.js
@@ -1,0 +1,10 @@
+import Nerv from 'nervjs'
+export default class Block extends Nerv.Component {
+  constructor (props) {
+    super(props)
+  }
+
+  render () {
+    return this.props.children
+  }
+}