+++ packages/taro-components/src/components/index.js
@@ -1,5 +1,6 @@
 import 'weui'
 export { default as View } from './view'
+export { default as Block } from './block'
 export { default as Image } from './image'
 export { default as Text } from './text'
 export { default as Switch } from './switch'