+++ packages/taro-components/src/components/swiper/swiper.js
@@ -222,7 +222,7 @@ class Swiper extends Nerv.Component {
       translate = -max
     } else {
       // default case
-      let changeV = this.isChangeSlide(translate, currentIndex) 
+      let changeV = this.isChangeSlide(translate, currentIndex)
       translate = changeV.translate
       currentIndex = changeV.currentIndex
     }
@@ -440,7 +440,7 @@ class Swiper extends Nerv.Component {
       width: this.state.wrapperWidth,
       height: this.state.wrapperHeight,
       transition: this.state.animating
-        ? `transform ${duration}ms ease-in-out;`
+        ? `transform ${duration}ms ease-in-out`
         : 'none',
       transform: `translate(${!vertical ? this.state.translate : 0}px, ${
         vertical ? this.state.translate : 0